+++ CHANGELOG.md
@@ -70,6 +70,7 @@ All notable changes to this project will be documented in this file.
 - [PR216] ffdec.sh - Support BSD version of sed in java version checks
 - Simple editor - adding a library item to the last frame adds a frame
 - Simple editor - white square on top left corner of the timeline
+- Bevel and Glow filters incorrect rendering
 
 ### Removed
 - Option to preview flash items via ActiveX component is no longer available.